Preview shot of landing page
In this preview shot, you're seeing the additional pages I am also working on and the landing page for my website. The picture inserted at the bottom is the flyer for the fashion show I am documenting. My goal for the website is to make it look like a gallery. I'm still trying to navigate what content I want where and the text to accompany it. I intend for the website to be image concentrated with some associated text, but the idea is to feel like you're learning about the process of the show as well as seeing the end product.